requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
